refactor: consolidate media metadata backfill processing
